bugfixes and cosmetic code changes
Some fixes here and there and small cosmetic changes in code mainly to have matlab stop complaining about stuff.

For some reason private functions InPolygon did not work for me, now everything is in the utils folder.

FeaturesUI: imfreehand is not recommended anymore, I tried to switch to drawfreehand. lokks like is working fine, we might consider changing also the one in SpikeImg

Block:
added SaveLoc.dir  in Paths to be consistent with the other fields. Changed the save function to reflect changes