Silhuette (cluster quality) update
Release notes:
This update introduces silhuette graphics in the features panel. This is a cluster quality metric that evaluates similarity between the points in the cluster and the cluster itself. Different distances metrics are implemented (at this stage the builtin ones) and selectable from a list (by double clicking on them). We can experiment a little and remove some of them in the future and add some new metrics/quality indexes. I spent a good chunk of yesterday trying to figure out how to use mutual information for this purpose. 

This metrics was chosen because it's local, that means that can give a measure for each cluster. I was thinking about introducing a second barplot on the right of this one (and a second list) for global measures, which give a shingle index to the whole clustering process. There is bazillion of them. WIP