Small functionality update
q button press in SpikeImg now reopens the FeatureUI(if closed).

Added the possibility to rotate the 3d feature plot bu pushing the rotate button. On my implementation for some weird reason this breaks the right click functionality on the 2D plot. To restore it, disable the rotation by clicking the button again.