updatePaths feature
Now it is possible to update the whole folder tree in the structure hierarchy. This is useful when something changes or you have to move the tank folder to a new location.

defaults.Block added the field FileNames. Fielnames can ideally be different to the corrispondent field in the channel structure. Is bad practice but you're free to do it.

getFolderTree calls an update of the Block feature. just in case you changed the default file. 

updateParams now handles also the case when 'block', 'animal', 'tank' or 'sort' are passed as inputs.

Added updatePaths functions to tank and animal