Remote Monitor
Added the monitor panel. The idea behind this is to spawn process in background (with the batch function) and have them write the state of the operation to a file. The file is read by the remote monitor and the progressbars are updated.
This works for process running in background on the local machine, I still have to figure out a way to get this working on a remote machine. 
The progressFiles (ie the files where the progress is stored) are created in the temp folder of the system. They are binary files with the first element being a int32 that encodes the total number of steps the operation is going to do. after that each time a step is done a uint8 1 is added to the files. By counting how many of those are added you get the progress till now. 

The progressMonitor relies on a timer to update the progressBars.