Pars Panel first commit
Pars Panel is now able to visualize all parameters in the Pars structure inside tank,block or animal. 

I slightly changed Block in  order to add a Pars public property that contains all the other parameters fields. This is a "tappullo", but I think this design is better and more versatile than having all the parameters structure living in some ad hoc properties of the block. We should probably switch to this.