MuLtinAnimal update
evt : readded jobComplete event that went missing for some reason
splitCompleted event added to trigger updates in the DashBoards

libs: splitmultianimalsUI this is the Gui for the multianimals interface. It takes a DashBoardObj as input during construction and communicates back and forth with it using mostly events.

utils : multiCallBackWrap is a wrapper to execute multiple functions during a callback

Block : changed all reference to ManyAnimals in MultiAnimals

reportProgress update : reportprogress now checks the last displayed characters to decide whether to update only the completed percentage or print everything back again. 
It also understands if it's ini worker environment or not, 

SplitMultiAnimals is defined both as a block and an animal method. In blocks it looks for the multiAnimal character and creates two children copy block stored in MultiAnimalsLinkedBlocks. 
In animals does the same, but with animals; 
If a TreeObj is passed as an input it applies the changes stored in the tree;

UpdatePaths now removes the old Block savefile.