Clean up Listener syntax
Many of the listeners were added without a reference to the listener object. I keep the listener object as properties of Tank, Animal, or Block and then ensure that during saveObj and delete that those property values are either transiently unassigned or deleted appropriately. In addition, during call to Tank.assignNULL or Animal.assignNULL, the 'Listener' property of the corresponding child object is the actual listener handle for that event (to avoid creating an array of event listeners with messy indexing and an ever-expanding array of deleted listener handles).