Move large methods to separate files
Block  -- Moved subsref to its own file. Ideally we will eventually turn this into a .p file or something like that so it is really hard to mess with because the bulk of the problems come from this thing.

Animal  --  Move subsref to its own file for same reason as Block (although this one is less troublesome at least). Move addChildBlock to its own file as well, since that was getting large.

Tank  --  Move subsref and addAnimal to own files.