Current status: multiAnimals not working
Right now it cycles through and deletes all the child blocks and the child "multi-animal" as well and doesn't reload anything. The "tree" interfaces do not populate. I think it's probably something in either Block or Animal but not sure what.

nigelColors  --  Docs formatting

DashBoard  --  Added SelectionIndex SetObservable property that works in combination with new features on nigelButton to prevent "Split" option from enabling on inappropriate nodes.

nigelButton  --  Added 'Enable' SetObservable property and changed the way the ButtonDownFcn works from being on a listener that updates the functions to having it be split into two properties: 'Fcn' and 'Fcn_Args' which are then called from a static 'ButtonDownFcn' assigned to the rectangle and text graphics. Added a 'highlight' edge to indicate that the button has been clicked, which stops glowing when button is released.

splitMultiAnimalsUI  -- Went through and commented on a bunch of stuff trying to understand how it works and add some basic documentation to it. Added property validation to different properties.

Did the same thing as for the libs class for both associated methods of nigeLab.Block and nigeLab.Animal as well.