Preliminary working version of doActions
`nigeLab.defaults.doActions`  --  Each pars field is the name of a `doMethod`. Field contents are cell arrays of the Fields that must evaluate ALL status flags to true in order for that `doMethod` to run correctly. 

`nigeLab.Block\checkActionIsValid`  --  If any of the requirements fail as described above (or if the `doMethod` name has not been configured yet, which is parsed dynamically from the method name), then it throws an error and puts a link in the Command Window to `nigeLab.defaults.doActions` so it can be configured properly. The error message also states which field failed as a necessary pre-processing step.

`doMethods`  --  Added `nigeLab.Block\checkActionIsValid();` as one line of code at the start of each of them. It should be added to any `doMethod` in the future.