doAutoClustering update final
doAutoClustering is now working on animal and block

when performed on an animal it stacks together all the spikes belonging to the same channel, before clustering. If channels are different between blocks, the clustering is done on the more restrictive set. The user is notified of the event.

doSD does not initialize Clusters and Sorted anymore