Feature: StatusChangedListener (#51)
* MAJOR BUGFIX and add getOperationField, update linkToData

nigeLab.Block/linkChannelsField and nigeLab.Block/intan2Block had a major bug in which blockObj.Channels(iCh).port_num was being used instead of .probe in the assignment of file names. This caused an error in nigeLab.Block/doReReference, which does use the correct .probe reference; it's recommended to extract any previously extracted data or at least check that file names correspond to correct probes.

Added ability for nigeLab.Block/linkToData to specify a 'field' argument so that calls to `linkToData` can run faster without having to cycle through all fields if it's not needed.

Add listener to nigeLab.DashBoard for `StatusChanged` event of .Tank property. This should dynamically update the "recap circle" table.

* Minor bugfixes

StatusChanged listener should be working as intended now.

* Add .gitattributes

For integration with git-lfs. This changes the .png, .jpg, .jpeg, .svg, and .mat files to "pointers" on the remote, so that the repo doesn't make a COPY of them each time there is a change. Should help reduce size of the repository.

* Feature: statusChangedListener tenatively complete

nigeLab.Block/updateStatus issues the 'StatusChanged' event and associated nigeLab.evt.statusChangedEventData eventData object with 'field', 'fieldType', 'key', 'status', and 'index' fields that relate the currently-updated status to the relevant field and key of a given block.

* Minor bugfix

Add drawnow; to updateStatus of nigelProgress object