Fix bug in computing completion % for doReReference
Added nigeLab.evt.barCleared event to delineate between `barStopped` vs when it is canceled by the red 'X'.

Added `joblistener` property to bar that is created when a job is added to the bar and listens for the job's destruction. That way if the job is destroyed via the Job Monitor it does not break the whole interface.