Working
Some weird errors if run at Animal or Tank level, but seemed like the extraction ran fine anyways so I think it might just be due to glitches with read/write due to OS hanging or something on the remote? Not sure, at any rate, further development should be done in a different branch.