Probe parsing based on BlockMask
parseProbes method of Animal is now updated to "re-scan" the child Blocks whenever the .BlockMask property of Animal is updated, so that for example if the "UnifyChildBlockMask" animalObj.Pars.Animal field is set to true (+defaults.Animal), then it will correctly unify the mask taking the new BlockMask into account. So if you have one day where there are only a few channels, this way you can remove that day using the BlockMask, and then it will unify the channel masking based on the remaining "good" recordings.