Add contextmenu enables, new menu options
Add contextmenu option for `Enable` (toggle .BlockMask for selected Blocks) and `Spike Sorting` (open the `Sort` interface using the selected Blocks).

The contextmenu now dynamically updates to gray out disabled `doMethods` and does not allow you to run the `Sort` class if the appropriate conditions are not met based on current selection. If any of the Blocks in the selection are "disabled" (Block.IsMasked, which is linked to Animal.BlockMask), then `doMethods` cannot be run. To reflect this, the `Enable` contextmenu item has the check next to it removed, and the selection highlighting changes to red on the ui tree.