feature/blockMask (#53)
* Add AnimalMask and BlockMask as well as loadobj inits

* Added property: Tank.AnimalMask -- Logical indexing vector to mask Animals.

* Added property: Animal.BlockMask -- Logical indexing vector to mask Blocks.

These will be `SetObservable` properties. For Animal, parseProbes should be invoked whenever the BlockMask is changed, as the "unified" probe mask (across all recording sessions, for example) depends upon which days are present. It may be desirable to "Mask" a recording where something went wrong with the probes.

* Modify +defaults.doActions and +defaults.nigelColors

*defaults.doActions now returns params struct with top-level same as before that is all fieldnames that correspond to `doMethods` (e.g. doRawExtraction); however, instead of these being cell arrays, they are now struct, which have the fields `required` (previous-version cell array) and `enabled` (logical field that can be set to false to "disable" the `doMethod` from nigeLab.libs.DashBoard interface)

*defaults.nigelColors - Added optional input 'help' or 'opts' to print out a list of current character input options corresponding to the colors that they represent to make it easier to remember what goes to what.

* Add evt.treeSelectionChanged

*evt.treeSelectionChanged  -  Event class that is issued whenever the `treeSelectionFcn` method of `nigeLab.libs.DashBoard` is called. Contains properties .Animal (handle array to each unique Animal object), .Block (handle array to each unique Block object), .Tank (handle to the tank), and .SelectionIndex (e.g. .Block = .Tank{.SelectionIndex}).

* Bugfix for KEY stuff, doActions

* Incorporate modifications to `checkActionIsValid` corresponding to changes in `+defaults.doActions`.

* `findByKey` method of Block and Animal had a bug where it was not correctly finding the key and instead would always point to the first array element. It now returns the correct block or animal. Also, instead of a key, the block or animal itself can be passed directly and a second output argument, the index within the larger array, can be obtained (useful for getting the matched element of a corresponding array when you want to find the thing that corresponds to a specific block in an array of blocks from a second array for example).

* Probe parsing based on BlockMask

parseProbes method of Animal is now updated to "re-scan" the child Blocks whenever the .BlockMask property of Animal is updated, so that for example if the "UnifyChildBlockMask" animalObj.Pars.Animal field is set to true (+defaults.Animal), then it will correctly unify the mask taking the new BlockMask into account. So if you have one day where there are only a few channels, this way you can remove that day using the BlockMask, and then it will unify the channel masking based on the remaining "good" recordings.

* Add contextmenu enables, new menu options

Add contextmenu option for `Enable` (toggle .BlockMask for selected Blocks) and `Spike Sorting` (open the `Sort` interface using the selected Blocks).

The contextmenu now dynamically updates to gray out disabled `doMethods` and does not allow you to run the `Sort` class if the appropriate conditions are not met based on current selection. If any of the Blocks in the selection are "disabled" (Block.IsMasked, which is linked to Animal.BlockMask), then `doMethods` cannot be run. To reflect this, the `Enable` contextmenu item has the check next to it removed, and the selection highlighting changes to red on the ui tree.

* Bugfix on AutoClustering

Changed so it does not update `Sorted` status any more.