Streams now initialize properly (#57)
MAIN "FEATURE" CHANGES:

* initChannelStruct : wasn't properly initializing Streams to array structs
* getStatus: Changed from 'any' to 'all' for Streams
* Add checks to reduce updateParams
* Add formatted print statement on load of individual objects.

Changed several things in the `init` and `doRawExtraction` files, which stems with adding the `.Samples` property to the `nigeLab.utils.signal` class. Now, each data stream ('Channels', 'Streams', or 'VidStreams') should all have the total number of samples parsed and associated with it via the `.signal` struct field in the corresponding property type.

Previously, there were issues due to trying to parse "Num[evaluated expression]Channels" variables before that were hard to diagnose. So this is in an attempt to eliminate that kind of confusion.

Bugfixes:
* Typos and updated variable references due to other changes made.

* Was getting weird java error from reportProgress on part of if statement that shouldn't be reached on remote. Tried changing checkForWorker() to default to true if the .OnRemote flag is true; was still getting weird instances in `reportProgress` where I was inside of a logical condition that I shouldn't be.

* It seems that using reportProgress on remote without setting 'OnRemote' flag causes java exception in getLastDispText() utility method. Setting `try` on it doesn't pass the java error. I think it's because it's on a remote worker that may not even have a Command Window, while trying to access the Command Window contents. Just changed it to modify the job tag directly. Anyways, just something to be aware of, that the `com` java stuff gets cranky when on remote.