Now is possible to spawn jobs to a local cluster
Local clusters are not managed by MJS but by CJS. A couple of tweaks were necessary to add support to this case:

added a flag IsParallel to the bar object in nigelProgress.
The completion of the task when ran on a local cluster is determined by the state of the job and the progress percentage.

Added a check in findgoodclusters. If the profiles are not found it falls back to the local one.

jobTag2Pct now supports CJS