Hotfixes
- buildWorkerConfigSccript
fixed some bugs in the  paths that were passed around. Now it should work both locally and remotely. Also, the wrapperfile is now positioned in the temp folder, to avoid cluttering the main repo.

- animal/init
fixed a bug that was picking up whatever file was present in an animal rawFile folder as a source file. Now it specifically looks for files with the filetype specified in the supportedFormats config.

- intan2Block
fixed a bug that was auomatically assigning to all the fields defined as 'Channels' the same number of channels as those of rawData. This is generally true but it can lead to errors when some additional recordings are turned off (e.g DC). When this happens it will also trigger 'TheWorstError'.
Time is now initialized to zeros. Not doing this was leading to an incorrect dimensioned time vector when RawExtraction was run multiple times on the same block. (due to appending instead of overwriting during initialization).
buffer.Dig.DigIo is initialized to empty struct. Failed initializatin was leading to an error down the line (It was looking for the fields of buffer.Dig.DigIO, which was not existing)
line 414, corrected a copyPasta typo.

- linkStreamsField
now checks for the existence of the field to be linked inside blockObj.Streams in order not to trigger any errors when the block tries to synch files that don't exists.

- list
Added some checks to avoid errors due to lack of metadata

- block/subsref
Added an exception vector for the shortcut .reference. blockObj.Time was looking for blockObj.Channels.Time leading to an error due to the lack of Time field in Channels (Time is a field of blockObj not Channels).