Update dev with incomplete VideoTools (#75)
* Bugfix

Some typos and minor stuff.

* Current State: scoreVideo nearly complete

* Changed `blockObj.getStreams` to return `nigeLab.libs.nigelStream` class object.
* Re-working interfaces to move away from confusing `listener` callback syntax, which seems to complicate and also slow down everything.
* Fixed issue with `getopt` that was causing some bugs
* scoreVideo loads from listbox correctly now.
* Multi-video now working: can click on "Time-Scroller" axes and it will automatically parse and jump to the correct video (which now actually loads quickly).
* Fixed some issues with `rollover` object
* Check out `nigeLab.defaults.nigelColors();` when called with no inputs!

* Current State: alignVideo nearly complete

* REMOVE CONFUSING LISTENER SYNTAX
  + Trying to stay away from so many listener callbacks in the future to reduce complexity.

* ADDED "TIME-SCROLLER"
  + Basically, merges functionality of scoreVideo and alignVideo into `scoreVideo` alone. Can now "jump" to frames by clicking in the axes denoting current frame time marker.
  + Allows clicking of streams to "drag" them and update neural offset.
  + Automatically parses the correct video from a list of videos based on what the "series time" is from within a given "camera source."
     - WIP: Make sure that non-consecutive (e.g. "extracted trials") videos work with this feature (should; may want to add graphical indicator of clickable/non-clickable regions to the axes).
  + WIP: Add ability to set individual-trial, individual-camera "fine-tuned" offsets.

* UPDATED COLORS
  + Updated `nigeLab.defaults.nigelColors()`
  + Moved `cubehelix` to `+utils`
  + Fixed some problems with `rollover` object

* ADDED BUFFERING
  + Can now left-click and drag on video image to select a (cropped) ROI on the video to zoom in and reduce the amount of data that is changed in the image frame when a new frame is read.
  + Unfortunately, the `readFrame` method of VideoReader still reads the entire video frame; however, the smaller size of the stored ROI is used to create a small buffer tensor that is indexed to allow searches of nearby frames to be more efficient.

Bugfixes

* MultiAnimals fixes

Many fixes and tweaks preformed to restore MultiAnimals capabilities

- {} indexing by ID
Dashboard and related objects have been updated to use the new {} indexing by ID. Now the Tree entries in the dashboard refer to the proper nigelObj using its unique ID. This simplifies a lot the deletion, addition and movements of the said objects.
Files changed: barCleared, barStarted, barStopped, DashBoard, nigelProgress, qOperations, remoteMonitor.

- Various fixes to make splitMultiAnimals current
Changed some stuff to make the branch current with the new nigelObj superclass organization. And some other small fixes.
File changed populateTree, splitMultiAnimalsUI, nigelObj, block/splitMultiAnimals

- Code reorganization to improve clarity and readability
File Changed: animal/splitMultiAnimals

* Hotfixes

- buildWorkerConfigSccript
fixed some bugs in the  paths that were passed around. Now it should work both locally and remotely. Also, the wrapperfile is now positioned in the temp folder, to avoid cluttering the main repo.

- animal/init
fixed a bug that was picking up whatever file was present in an animal rawFile folder as a source file. Now it specifically looks for files with the filetype specified in the supportedFormats config.

- intan2Block
fixed a bug that was auomatically assigning to all the fields defined as 'Channels' the same number of channels as those of rawData. This is generally true but it can lead to errors when some additional recordings are turned off (e.g DC). When this happens it will also trigger 'TheWorstError'.
Time is now initialized to zeros. Not doing this was leading to an incorrect dimensioned time vector when RawExtraction was run multiple times on the same block. (due to appending instead of overwriting during initialization).
buffer.Dig.DigIo is initialized to empty struct. Failed initializatin was leading to an error down the line (It was looking for the fields of buffer.Dig.DigIO, which was not existing)
line 414, corrected a copyPasta typo.

- linkStreamsField
now checks for the existence of the field to be linked inside blockObj.Streams in order not to trigger any errors when the block tries to synch files that don't exists.

- list
Added some checks to avoid errors due to lack of metadata

- block/subsref
Added an exception vector for the shortcut .reference. blockObj.Time was looking for blockObj.Channels.Time leading to an error due to the lack of Time field in Channels (Time is a field of blockObj not Channels).

* Bugfixes

* Fix one line that got bumped out of place during merge somehow.

* Added `nigeLab.libs.nigelCamera`

* Small helper class to associate a series of videos from the same camera with each other.
* Ensures that when a particular time is selected, the correct corresponding video index is selected, facilitating things like skipping around on the TimeAxes when there are multiple videos.

* Video Alignment working

* Now updates the trial marker timestamps when alignment is updated.
  + WIP: Automatically export trial "chunk" videos.

* Remove VideoReader

* Removed deprecated VideoReader modifications

Co-authored-by: Nabarb <Nabarb@users.noreply.github.com>