nigelObj improvements
Bugfix on addchild. Parent is now assigned properly when multiple chidren are added at once.

Bugfix on reload. Reloading transient properties was causing their value to be overwritten by `null` or empty. Now they are skipped during reload.

Bugfix on checkAnimalsForClones. It wasn't warking at all. I ended up rewriting it simplifying it a bit.

Added lookForData function. While loading if no children are added to the loaded obj (because none ere found) the user is prompted to select the correct data folder. This triggers a path update for the nigelObj (and subsequently loaded children).