various fixes
Various fixes.

the demo_nigelab now works to the end.

A small tweak in nigelObj to make sure that the data are correctly retrieved when moved.

Fixed a very rare fix in the artifact detection step.

getNigelPath was returning a hardcoded path to kumc internal directories when run in unix.