Bugfix/sortfix (#84)
* Sort bugfix

Overall minor fixes to get Sort interface to work properly

- Recluster kmeans now is done on the feature space obtained by the wavelet decomposition. I'm planning to include some toggles to toggle between feature spaces.

SpikeImage Added some other shortcuts. Will add documentation in the wiki soon.

* Diskdata tweaks

Proposed diskdata  tweaks. Should improve performance a little. It is taking a really long time to save and load all the required data during sorting.

Major tweak: hd5write/read calls reduced by writing/loading files in chunks instead of column by column.

Old version is commented out instead of removing it to make it easily reversible.

* UImenuclick_sort fix

Modified uimenuclick to let the sortObj figure out what to do with the input arg.
Also added a small error if the user tries to sort more then one animal or the tank.

* Dashboard opening fix

Ported changes from multiAnimals branch to fix an issue with the dashboard not opening correctly

* fix nargout in diskdata

fixed weird bug that was causing subsref to crash due to a wrong number returned by nargout. This is not a definitive fix. Also, is happening only on my laptop (Matlab 2017a and b) but not on my workstation (2017a and 2019b). Will investigate further.

* Revert commit for demo

* various fixes

Various fixes.

the demo_nigelab now works to the end.

A small tweak in nigelObj to make sure that the data are correctly retrieved when moved.

Fixed a very rare fix in the artifact detection step.

getNigelPath was returning a hardcoded path to kumc internal directories when run in unix.