hotfix
hotfix collection

solved a bug that was causing the Sort interface to change channel when unconfirmed changes were present, regardless the user answer to the message prompt.

Solved a bug in Dashboard that was preventing the "status" figure to update properly after a task was completed.

Silenced some warnings about java in matlab 2019b and later. I reserve the pleasure of dealing with this to the future me.