Spike Detection new features
doSD was essentially rewritten. All the SD algorithms were moved to their own functions in the block private folder. The same was done for feature extraction and artefact rejection.

This allows for a very flexible configuration of the pipeline. The SD of choice can be specified in the defaults SD file and the ad hoc parameters can be customized in the default/SD folder.
this is similar to what was done with the AutoClustering method.

A shirt wiki entry will follow with the style specifications and naming convention of the files.