Feature/new sort (#88)
* Sort new features

Added a pletora of features to the Sort interface.
Fixed and improved the way the SortFiles are saved and loaded, (getClus getSort). Fixed small bugs here and there,

HighDimsUi now has the same interface language as dashboard (with nigelpanels etc). A new button was added to find the most convenient projection according to LDA. When pressed moves the plane visualized in featuresUI in the position that best separates the selected cluster.
Thanks DataHigh!

ginput
Changed the cross-hair color to improve visibility.

multicallbackwrap wasn't really working. fixed.

NigelButton
fixed a small bug that was overwriting WindowButtonUpFnc of the figure it was put on. Now it uses multiCallbackWrap to avoid overwriting.
FaturesUI
The 2D graph now cna be zoomed and panned. Use the MouseWheel to zoom. MouseWheel + alt or control to pan horizontally and vertically.

Saving performance greatly improved by saving only the clusters that actually need saving. Each channels starts with a UnsavedChanges target to false and a ConfirmedChanges to false. When a change is done the UnsavedChanges goes to true. When confirmed, ConfirmedChanges goes to true.
SpikeImage; SortUi;

* Fixed zooming in FeaturesUI

Now the two graphs have linked axes that change and shift together.

Also, navigation is now homogeneous between the two figures

* Fixed bug in Sort/savedata

fixed a bug in savedata. Program was crashing when a channel was turned off in mask

* performance update

Small performance update in plotting and updating spike profiles and features.

* Added pca and random projection

Added the possibility to project data onto the plane of max variance as well as to a random projection plane.

This is accessible from the HighDimsUi