Wait message added
Added a message prompting the user to wait for the program to finish. While the message is displayed the UI is blocked. This prevents the user to get nigel-level impatient and mess up everything.

The resample button now also works as intended.

ToDo add the export feature to directly export the selected parameters to the blockObj.Pars.SD struct