Feature/new sd (#92)
# New Spike Detection
Added by: Federico (FB)

## Features
### Change in Structure
`doSD` has been essentially re-written. 
* All the SD algorithms were moved to their own functions in the block private folder. 
* The same was done for feature extraction and artifact rejection.

#### Motivation
We want the pipeline to be more flexible, particularly in the spike detection step where many labs or users have their own preferred method. Reorganizing the structure of `doSD` in this way allows the specification of SD choices both in the `defaults.SD` file, as well as on an ad hoc basis through a user interface.

#### To-Do
* A short wiki entry will follow with the style specifications and naming convention of the files.

### Spike Detection Algorithm Updates
* Revised and fixed `AdaptThresh`, `HardThresh`, `SNEO`, and `TIFCO`.
* Fixed associated `SD` parameters.

### `configSD`
Added a GUI to configure the SD parameters. 

#### Strategy
It randomly extracts 1 minute of data from the recording, providing a graphical interface that displays the data and any spikes detected as the end-user tries different parameter configurations until they are satisfied that the current parameter settings can detect spikes well in the data.  and you can try different parameters on the extracted data.
* Artifact rejection parameters can be selected in the same way.
* The "current channel" can be changed by the end-user.