TDT minimal functionality update 2
Found an insidious bug in nigelObj. The system was using a function to initialize some parameters. The problem is matlab caches the class file the first time it is executed, resulting in those values not updating and retaining the default value of the first class run.
     -> This was solved by simply removing the initialization and putting it back at the beginning of the constructor function

Other small fixes for tdt compatibility