Stim suppression method
Added suppressStmi method to block. This returns a raw signal cleaned off the stim Artefact.

Methods can be added to the private Block folder as long as the correspondent pars are added in the default folder. By default it is active in Filt and LFP.

Also added a notch filter to cut the main power interference