bugfixes
nigelCamera fixed an error in the order of operations during initialization that was leading to the cam Obj not having a time vector

fixed an error issued when trying to run the extractMean from an inactive source. Now this operation is possible.

 VidScorer fixed an error issued when running the extract signal from mean option in the videostream context menu

fixed a typo in the delete stream function

doEventDetection
added a force parameter to force the operation without going through the dialog

initVideos
added a small check to make sure that videos are not added twice to the camera obj