Functionality update to video suite
Nigelcamera: VideoReader and TS props are now dependent and set upon request when empty. This speeds up loading of Tank a lot when many videos are linked.
When a the VideoReader is reloaded Meta is updated keeping the old values unchanged when not present in the new Meta struct. This also ensures that the synch and stretch pars are not overwritten.
Other small fixes here and there.

simpleVideoReader.cpp : Fixed an issue that was causing the program to fill the msBuffer (the vector keeping track of ms time associated w each frame) with nans when frames are corrupted/unreadable. This was causing the msBuffer not to be aligned to frameBuffer.
Fixed some issues with the seek function. Now works as properly.
Other small fixes around.

VidScorer: fixed some glitches here and there. Now the main camera is set at construction. The speed is also reset for the main camera at construction.
Each event lights up when the corresponding frame is shown (only as a result of frameB/F and seek. Nor during play).
Some minor usability fixes.

add2struct was breaking the update of camera's Meta. fixed.

orderGoProVideos was broken. fixed.

doEventDetection now checks for duplicates and removes them,