Updates to deal with new pars struct
Updated the initialization of blocks to create the correct structure of pars since the tank initialization. Updated the functions that performs the stimsuppression, SD and Sorting to manage the new struct of pars. Updated the updatePars function so that it is able to give back the correct struct of paramaters if needed.