fix initChannels
Block:
Fixed initialization with one pars per channels.
Dynamic naming of folders following methods was removed since it was already broken and was introducing some difficulties.

NumChannels had a set method that was not supposed to be there.  Removing it solved an initialization error.

initChannles:
removed the NumChannels set line that would not work at this point.

defaults/Block:
removed support for dynamic naming