hotFix AdSWTTEO
The previous version of SD_AdSWTTEO was wrong. Added the possibilty to parametrize also the k parameter resambling the number of samples to take into account for the  computation of the TEO. There is a new custom function (mymovquant) to compute the moving quantile and the AdSWTTEO now works similarly to the PTSD (also with the overshoot secondary window). The step parameter of the previous versoin is dropped.